requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
discord.py==2.1.0
frozenlist==1.3.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.2.0
multidict==6.0.4
nltk==3.8.1
numpy==1.24.1
pycodestyle==2.10.0
python-dotenv==0.21.0
regex==2022.10.31
requests==2.28.1
rfc3986==1.5.0
scikit-learn==1.2.0
scipy==1.9.3
sklearn==0.0.post1
sniffio==1.3.0
threadpoolctl==3.1.0
tqdm==4.64.1
urllib3==1.26.13
yarl==1.8.2
